requirements.txt
Flask==3.0.3
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.0.5
Flask-Smorest==0.44.0
marshmallow==3.20.1
